21 online game   component

+++ TTT-backend/src/main/java/com/codecool/tttbackend/dao/UserDAO.java
@@ -6,7 +6,7 @@ public abstract interface UserDAO {
 
     User findByUsername(String username);
     User findByEmail(String email);
-    User findUserById(Long id);
+    User findUserById(int id);
     void addNewUser(User user);
     void updateUser (User user);
     void deleteUserById(int id);
